Split code into readable modules
