+++ lib/transporter.js
@@ -1,0 +1,15 @@
+
+const nodemailer = require('nodemailer');
+
+// Setup nodemailer
+const transporter = nodemailer.createTransport({
+    host: process.env.EMAIL_HOST,
+    port: process.env.EMAIL_PORT,
+    secure: process.env.EMAIL_SECURE === "true",
+    auth: {
+        user: process.env.EMAIL_USER,
+        pass: process.env.EMAIL_PASS
+    }
+});
+
+module.exports = transporter;
