+++ routes/root.js
@@ -1,9 +1,7 @@
 var express = require('express');
 var router = express.Router();
 
-router.get('/', (req, res) => {
-    return showServiceRunning(res);
-});
+router.get('/', (req, res) => showServiceRunning(res));
 
 // Show message that service is running
 function showServiceRunning(res) {
